[View][Unittest] enabled to automatically create today record. And split the file for test into 2 files.
